store: Move find_transaction_receipts_in_block fn to Storage methods
