+++ store/postgres/src/transaction_receipt.rs
@@ -1,73 +1,12 @@
-use diesel::{
-    pg::PgConnection,
-    prelude::*,
-    sql_query,
-    sql_types::{Binary, Nullable, Text},
-};
+use diesel::sql_types::{Binary, Nullable};
 use diesel_derives::QueryableByName;
-use graph::prelude::web3::types::H256;
+use graph::prelude::transaction_receipt::LightTransactionReceipt;
 use itertools::Itertools;
 use std::convert::TryFrom;
 
-use graph::prelude::transaction_receipt::LightTransactionReceipt;
-
-use crate::ETHEREUM_BLOCKS_TABLE_NAME;
-
-/// Queries the database for all the transaction receipts in a given block range.
-pub fn find_transaction_receipts_in_block(
-    conn: &PgConnection,
-    blocks_table_name: &str,
-    block_hash: H256,
-) -> anyhow::Result<Vec<LightTransactionReceipt>> {
-    let query = sql_query(format!(
-        "
-select
-    ethereum_hex_to_bytea(receipt ->> 'transactionHash') as transaction_hash,
-    ethereum_hex_to_bytea(receipt ->> 'transactionIndex') as transaction_index,
-    ethereum_hex_to_bytea(receipt ->> 'blockHash') as block_hash,
-    ethereum_hex_to_bytea(receipt ->> 'blockNumber') as block_number,
-    ethereum_hex_to_bytea(receipt ->> 'gasUsed') as gas_used,
-    ethereum_hex_to_bytea(receipt ->> 'status') as status
-from (
-    select
-        jsonb_array_elements(data -> 'transaction_receipts') as receipt
-    from
-        {blocks_table}
-    where hash = $1) as temp;
-",
-        blocks_table = blocks_table_name
-    ));
-
-    let query_results: Result<Vec<RawTransactionReceipt>, diesel::result::Error> = {
-        // The `hash` column has different types between the `public.ethereum_blocks` and the
-        // `chain*.blocks` tables, so we must check which one is being queried to bind the
-        // `block_hash` parameter to the correct type
-        if blocks_table_name == ETHEREUM_BLOCKS_TABLE_NAME {
-            query
-                .bind::<Text, _>(format!("{:x}", block_hash))
-                .get_results(conn)
-        } else {
-            query
-                .bind::<Binary, _>(block_hash.as_bytes())
-                .get_results(conn)
-        }
-    };
-
-    query_results
-        .map_err(|error| {
-            anyhow::anyhow!(
-                "Error fetching transaction receipt from database: {}",
-                error
-            )
-        })?
-        .into_iter()
-        .map(LightTransactionReceipt::try_from)
-        .collect()
-}
-
 /// Type that comes straight out of a SQL query
 #[derive(QueryableByName)]
-struct RawTransactionReceipt {
+pub(crate) struct RawTransactionReceipt {
     #[sql_type = "Binary"]
     transaction_hash: Vec<u8>,
     #[sql_type = "Binary"]
